requirements.txt
python-osc==1.4.2
python-rtmidi==0.5b1
